Add files via upload
Our metric for optimization is the number of misclassification/errors from each model. We only use a small amount of data to train(600 samples total and 200 for each class) and validate(150 samples total and 50 for each class) our model. 

The beauty of transfer learning is that we can use the weights and parameters these CNNs have learned over hours of training over millions of images. We can use the initial layers for generic feature extraction, or all layers except the fully connected classifier layers or even the full CNN to make predictions as we have in this notebook.